Completes the FEM course
