+++ src/shapes/mod.rs
@@ -2,3 +2,4 @@ pub mod area;
 pub mod circle;
 pub mod collisions;
 pub mod rect;
+pub mod shape;
