+++ src/shapes_input
@@ -1,0 +1,5 @@
+rect 1 2 3 4
+circle 0 0 2
+circle 0 0 1
+rect 2 1 4 5
+circle 1 1 2
